+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Changed
 
+- [Text](/#/Components?id=text) component
 - Add a `shrink` option to [Dropdown](/#/Components?id=dropdown) to not have a min-width.
 - Fix heights of different [Badge](/#/Components?id=badge) component variants.
 - Made [Field](/#/Components?id=field) error messages smaller