+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix heights of different [Badge](/#/Components?id=badge) component variants.
 - Made [Field](/#/Components?id=field) error messages smaller
 - Remove [Field](/#/Components?id=field) error messages altogether when `reserveSpaceForError` is false and there's no error.
+- Make [Icon](/#/Components?id=icon) render it's `width` and `height` props in CSS instead of `svg` attributes.
 
 ## [5.6.0] - 2019-05-01